Factor out `stabilize()` from K8s for reuse in Consul
