+++ k8s/src/main/scala/io/buoyant/k8s/package.scala
@@ -26,76 +26,4 @@ package object k8s {
     new JsonStreamParser(mapper)
   }
 
-  type VarUp[T] = Var[T] with Updatable[T]
-
-  /**
-   * We can stabilize this by changing the type to Var[Option[Var[T]]].
-   * If this Option changes from None to Some or vice versa, the outer Var will
-   * update.  If the value contained in the Some changes, only the inner Var
-   * will update.
-   */
-  def stabilize[T](unstable: Var[Option[T]]): Var[Option[Var[T]]] = {
-    val init = unstable.sample().map(Var(_))
-    Var.async[Option[VarUp[T]]](init) { update =>
-      // the current inner Var, null if the outer Var is None
-      @volatile var current: VarUp[T] = null
-      @volatile var exists = false
-      unstable.changes.respond {
-        case Some(t) if current == null =>
-          // T created
-          exists = true
-          current = Var(t)
-          update() = Some(current)
-        case Some(t) if !exists =>
-          // T re-created
-          exists = true
-          current() = t
-          update() = Some(current)
-        case Some(t) =>
-          // T modified
-          current() = t
-        case None =>
-          // T deleted
-          exists = false
-          update() = None
-      }
-    }
-  }
-
-  // NOTE: we may want to consider naming this function, since there is also
-  //       a method `Activity.stabilize()` on `Activity` which has completely
-  //       different behaviour. This seems confusing.
-  //
-  //       N.B. that if we renamed this function, we probably would want to
-  //       also rename the corresponding `stabilize(Var)` function that was
-  //       the source of this name in the first place...
-  def stabilize[T](unstable: Activity[Option[T]]): Activity[Option[Var[T]]] = {
-    val inner = Var.async[Activity.State[Option[VarUp[T]]]](Activity.Pending) { update =>
-      // the current inner Var, null if the outer Var is None
-      @volatile var current: VarUp[T] = null
-      @volatile var exists = false
-      unstable.states.respond {
-        case Activity.Ok(Some(t)) if current == null =>
-          // T created
-          current = Var(t)
-          exists = true
-          update() = Activity.Ok(Some(current))
-        case Activity.Ok(Some(t)) if !exists =>
-          // T recreated
-          exists = true
-          current() = t
-          update() = Activity.Ok(Some(current))
-        case Activity.Ok(Some(t)) =>
-          // T modified
-          current() = t
-        case Activity.Ok(None) =>
-          // T deleted
-          exists = false
-          update() = Activity.Ok(None)
-        case Activity.Pending =>
-          update() = Activity.Pending
-      }
-    }
-    Activity(inner)
-  }
 }
